[VCDA-852-GetOvdcResourcePool]
