+++ container_service_extension/broker.py
@@ -41,16 +41,15 @@ from container_service_extension.exceptions import NFSNodeCreationError
 from container_service_extension.exceptions import NodeCreationError
 from container_service_extension.exceptions import WorkerNodeCreationError
 from container_service_extension.logger import SERVER_LOGGER as LOGGER
+from container_service_extension.utils import ACCEPTED
 from container_service_extension.utils import ERROR_DESCRIPTION
 from container_service_extension.utils import ERROR_MESSAGE
 from container_service_extension.utils import ERROR_STACKTRACE
+from container_service_extension.utils import OK
 from container_service_extension.utils import SYSTEM_ORG_NAME
 from container_service_extension.utils import error_to_json
+from container_service_extension.utils import exception_handler
 
-OK = 200
-CREATED = 201
-ACCEPTED = 202
-INTERNAL_SERVER_ERROR = 500
 
 OP_CREATE_CLUSTER = 'create_cluster'
 OP_DELETE_CLUSTER = 'delete_cluster'
@@ -113,32 +112,6 @@ def rollback(func):
             except Exception as err:
                 LOGGER.error('Failed to rollback node creation:%s', str(err))
     return wrapper
-
-
-def exception_handler(func):
-    """ This function is used as decorator, executes the function that is passed as argument.
-    returns exactly what the passed function returns.
-
-    If there is any exception, returns new dictionary with keys status code and body.
-
-    NOTE: This decorator should be applied only on those functions that constructs the final
-    HTTP responses and also needs exception handler as additional behaviour.
-
-    :param func: original function that needs to be executed
-
-    :return: reference to the function that executes the passed function 'func'
-    """
-    @functools.wraps(func)
-    def exception_handler_wrapper(*args, **kwargs):
-        result = {}
-        try:
-            result = func(*args, **kwargs)
-        except Exception as err:
-            result['status_code'] = INTERNAL_SERVER_ERROR
-            result['body'] = error_to_json(err)
-            LOGGER.error(traceback.format_exc())
-        return result
-    return exception_handler_wrapper
 
 
 def task_callback(task):
@@ -558,7 +531,6 @@ class DefaultBroker(threading.Thread):
         result = {}
         result['body'] = {}
         LOGGER.debug('about to delete cluster with name: %s' % body['name'])
-        result['status_code'] = INTERNAL_SERVER_ERROR
 
         self.cluster_name = body['name']
         self.tenant_info = self._connect_tenant(headers)
